This closes issue #1 Hi @
eddieysong

+++ app/src/main/java/ca/aagavin/hospitalapp/DAO/CommonDAO.java
@@ -59,7 +59,7 @@ public class CommonDAO {
 
         cursor.moveToFirst();
         
-        doctor.setDoctorId(cursor.getLong(0));
+        doctor.setDoctorId(cursor.getInt(0));
         doctor.setFirstname(cursor.getString(1));
         doctor.setLastname(cursor.getString(2));
         doctor.setDepartment(cursor.getString(3));
@@ -77,7 +77,7 @@ public class CommonDAO {
 
         cursor.getCount();
 
-        doctor.setDoctorId(cursor.getLong(0));
+        doctor.setDoctorId(cursor.getInt(0));
         doctor.setFirstname(cursor.getString(1));
         doctor.setLastname(cursor.getString(2));
         doctor.setDepartment(cursor.getString(3));
